Refactor qildim
Refactor and add new code generation feature

Removed `CodeGenerator` class and related methods/files.
Replaced with new `Generator` class in `Program.cs`.
Added `Generator.cs` for generating service and controller classes.
Implemented `ControllerGenerator.cs` for controller code generation.
Added `IServiceGenerator.cs` for service interface generation.
Implemented `ServiceGenerator.cs` for service implementation.
Introduced helper methods for various code generation tasks.
Enhanced code generation for Entity Framework Core projects.
