+++ src/EntityCore.Tools/Services/ServiceGenerator.cs
@@ -1,0 +1,172 @@
+﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.EntityFrameworkCore;
+using System.Reflection;
+using Microsoft.CodeAnalysis;
+
+namespace EntityCore.Tools
+{
+    public partial class Generator
+    {
+        private string GenerateServiceImplementationCode(Assembly assembly, Type entityType, string? dbContextName)
+        {
+            string entityName = entityType.Name;
+
+            Type? dbContextType = null;
+
+            if (dbContextName is not null)
+            {
+                dbContextType = assembly.GetTypes()
+                                        .FirstOrDefault(t => typeof(DbContext).IsAssignableFrom(t)
+                                            && t.IsClass
+                                            && !t.IsAbstract
+                                            && t.Name == dbContextName
+                                        );
+            }
+            else
+            {
+                var dbContextTypes = assembly.GetTypes()
+                                             .Where(t => typeof(DbContext).IsAssignableFrom(t)
+                                                && t.IsClass
+                                                && !t.IsAbstract
+                                             );
+
+                if (dbContextTypes.Count() == 1)
+                    dbContextType = dbContextTypes.First();
+                else if (dbContextTypes.Count() > 1)
+                    throw new InvalidOperationException("Multiple DbContexts found in the specified assembly. Please choose DbContext name. ex: --context <DbContextName>");
+            }
+
+            if (dbContextType is null)
+                throw new InvalidOperationException("DbContext not found in the specified assembly.");
+
+            var primaryKey = FindKeyProperty(entityType);
+
+            var classDeclaration = GetClassDeclaration(entityName, primaryKey, dbContextType);
+
+            var namespaceDeclaration = SyntaxFactory.NamespaceDeclaration(SyntaxFactory.ParseName($"Services.{entityName}s"))
+                .AddMembers(classDeclaration);
+
+            CompilationUnitSyntax syntaxTree = GenerateUsings(namespaceDeclaration, dbContextType, entityType);
+
+            return syntaxTree.NormalizeWhitespace()
+                .ToFullString();
+        }
+
+        private ClassDeclarationSyntax GetClassDeclaration(string entityName, PropertyInfo primaryKey, Type dbContextType)
+        {
+            var dbContextVariableName = dbContextType.Name.GenerateFieldNameWithUnderscore();
+
+            var classDeclaration = SyntaxFactory.ClassDeclaration($"{entityName}sService")
+                        .AddBaseListTypes(SyntaxFactory.SimpleBaseType(SyntaxFactory.ParseTypeName($"I{entityName}sService")))
+                        .AddModifiers(SyntaxFactory.Token(SyntaxKind.PublicKeyword))
+                        .AddMembers(
+                            // fields
+                            SyntaxFactory.FieldDeclaration(
+                                SyntaxFactory.VariableDeclaration(
+                                    SyntaxFactory.ParseTypeName(dbContextType.Name))
+                                .AddVariables(SyntaxFactory.VariableDeclarator(dbContextVariableName)))
+                                .AddModifiers(
+                                    SyntaxFactory.Token(SyntaxKind.PrivateKeyword),   // private 
+                                    SyntaxFactory.Token(SyntaxKind.ReadOnlyKeyword)), // readonly
+                                                                                      //constructors
+                            SyntaxFactory.ConstructorDeclaration($"{entityName}sService")
+                                .AddModifiers(SyntaxFactory.Token(SyntaxKind.PublicKeyword))
+                                .AddParameterListParameters(SyntaxFactory.Parameter(SyntaxFactory.Identifier(dbContextType.Name.GenerateFieldName()))
+                                    .WithType(SyntaxFactory.ParseTypeName(dbContextType.Name)))
+                                .WithBody(SyntaxFactory.Block(SyntaxFactory.SingletonList<StatementSyntax>(
+                                    SyntaxFactory.ExpressionStatement(SyntaxFactory.ParseExpression($"{dbContextVariableName} = {dbContextType.Name.GenerateFieldName()}"))
+                                ))),
+
+                            // methods
+                            GenerateAddMethodImplementation(entityName, dbContextVariableName),
+                            GenerateGetAllMethodImplementation(entityName, dbContextVariableName),
+                            GenerateGetByIdMethodImplementation(entityName, dbContextVariableName, primaryKey),
+                            GenerateUpdateMethodImplementation(entityName, dbContextVariableName, primaryKey),
+                            GenerateDeleteMethodImplementation(entityName, dbContextVariableName, primaryKey)
+                        );
+
+            return classDeclaration;
+        }
+
+        private MethodDeclarationSyntax GenerateAddMethodImplementation(string entityName, string dbContextVariableName)
+        {
+            return SyntaxFactory.MethodDeclaration(SyntaxFactory.GenericName(SyntaxFactory.Identifier("Task"))
+                                .AddTypeArgumentListArguments(SyntaxFactory.ParseTypeName(entityName)), "AddAsync")
+                                .AddModifiers(
+                                    SyntaxFactory.Token(SyntaxKind.PublicKeyword), // public
+                                    SyntaxFactory.Token(SyntaxKind.AsyncKeyword))  // async
+                                .AddParameterListParameters(SyntaxFactory.Parameter(SyntaxFactory.Identifier("entity"))
+                                    .WithType(SyntaxFactory.ParseTypeName(entityName)))
+                                .WithBody(SyntaxFactory.Block(
+                                    SyntaxFactory.ParseStatement($"var entry = await {dbContextVariableName}.Set<{entityName}>().AddAsync(entity);"),
+                                    SyntaxFactory.ParseStatement($"await {dbContextVariableName}.SaveChangesAsync();"),
+                                    SyntaxFactory.ParseStatement($"return entry.Entity;")
+                                ));
+        }
+
+        private MethodDeclarationSyntax GenerateGetAllMethodImplementation(string entityName, string dbContextVariableName)
+        {
+            return SyntaxFactory.MethodDeclaration(SyntaxFactory.GenericName(SyntaxFactory.Identifier("Task"))
+                                .AddTypeArgumentListArguments(SyntaxFactory.ParseTypeName($"List<{entityName}>")), "GetAllAsync")
+                                .AddModifiers(
+                                    SyntaxFactory.Token(SyntaxKind.PublicKeyword), // public
+                                    SyntaxFactory.Token(SyntaxKind.AsyncKeyword))  // async
+                                .WithBody(SyntaxFactory.Block(
+                                    SyntaxFactory.ParseStatement($"return await {dbContextVariableName}.Set<{entityName}>().ToListAsync();")
+                                ));
+        }
+
+        private MethodDeclarationSyntax GenerateGetByIdMethodImplementation(string entityName, string dbContextVariableName, PropertyInfo primaryKey)
+        {
+            return SyntaxFactory.MethodDeclaration(SyntaxFactory.GenericName(SyntaxFactory.Identifier("Task"))
+                                .AddTypeArgumentListArguments(SyntaxFactory.ParseTypeName(entityName)), "GetByIdAsync")
+                                .AddModifiers(
+                                    SyntaxFactory.Token(SyntaxKind.PublicKeyword), // public
+                                    SyntaxFactory.Token(SyntaxKind.AsyncKeyword))  // async
+                                .AddParameterListParameters(SyntaxFactory.Parameter(SyntaxFactory.Identifier("id"))
+                                    .WithType(SyntaxFactory.ParseTypeName(primaryKey.PropertyType.ToCSharpTypeName())))
+                                .WithBody(SyntaxFactory.Block(
+                                    SyntaxFactory.ParseStatement($"var entity = await {dbContextVariableName}.Set<{entityName}>().FirstOrDefaultAsync(x => x.{primaryKey.Name} == id);"),
+                                    SyntaxFactory.ParseStatement($"if (entity == null) throw new InvalidOperationException($\"{entityName} with Id {{id}} not found.\");"),
+                                    SyntaxFactory.ParseStatement($"return entity;")
+                                ));
+        }
+
+        private MethodDeclarationSyntax GenerateUpdateMethodImplementation(string entityName, string dbContextVariableName, PropertyInfo primaryKey)
+        {
+            return SyntaxFactory.MethodDeclaration(SyntaxFactory.GenericName(SyntaxFactory.Identifier("Task"))
+                                .AddTypeArgumentListArguments(SyntaxFactory.ParseTypeName(entityName)), "UpdateAsync")
+                                .AddModifiers(
+                                    SyntaxFactory.Token(SyntaxKind.PublicKeyword), // public
+                                    SyntaxFactory.Token(SyntaxKind.AsyncKeyword))  // async
+                                .AddParameterListParameters(SyntaxFactory.Parameter(SyntaxFactory.Identifier("id"))
+                                    .WithType(SyntaxFactory.ParseTypeName(primaryKey.PropertyType.ToCSharpTypeName())))
+                                .AddParameterListParameters(SyntaxFactory.Parameter(SyntaxFactory.Identifier("entity"))
+                                    .WithType(SyntaxFactory.ParseTypeName(entityName)))
+                                .WithBody(SyntaxFactory.Block(
+                                    SyntaxFactory.ParseStatement($"var entry = {dbContextVariableName}.Set<{entityName}>().Update(entity);"),
+                                    SyntaxFactory.ParseStatement($"await {dbContextVariableName}.SaveChangesAsync();"),
+                                    SyntaxFactory.ParseStatement($"return entry.Entity;")
+                                ));
+        }
+
+        private MethodDeclarationSyntax GenerateDeleteMethodImplementation(string entityName, string dbContextVariableName, PropertyInfo primaryKey)
+        {
+            return SyntaxFactory.MethodDeclaration(SyntaxFactory.GenericName(SyntaxFactory.Identifier("Task"))
+                                .AddTypeArgumentListArguments(SyntaxFactory.ParseTypeName(entityName)), "DeleteAsync")
+                                .AddModifiers(
+                                    SyntaxFactory.Token(SyntaxKind.PublicKeyword), // public
+                                    SyntaxFactory.Token(SyntaxKind.AsyncKeyword))  // async
+                                .AddParameterListParameters(SyntaxFactory.Parameter(SyntaxFactory.Identifier("id"))
+                                    .WithType(SyntaxFactory.ParseTypeName(primaryKey.PropertyType.ToCSharpTypeName())))
+                                .WithBody(SyntaxFactory.Block(
+                                    SyntaxFactory.ParseStatement($"var entity = await {dbContextVariableName}.Set<{entityName}>().FirstOrDefaultAsync(x => x.{primaryKey.Name} == id);"),
+                                    SyntaxFactory.ParseStatement($"if (entity == null) throw new InvalidOperationException($\"{entityName} with {{id}} not found.\");"),
+                                    SyntaxFactory.ParseStatement($"var entry = {dbContextVariableName}.Set<{entityName}>().Remove(entity);"),
+                                    SyntaxFactory.ParseStatement($"await {dbContextVariableName}.SaveChangesAsync();"),
+                                    SyntaxFactory.ParseStatement($"return entry.Entity;")
+                                ));
+        }
+    }
+}
